Revert "tk10"
