+++ app/src/main/java/br/com/ecocycle/ecocycleapp/MainActivity.kt
@@ -3,32 +3,12 @@ package br.com.ecocycle.ecocycleapp
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.annotation.ColorRes
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import br.com.ecocycle.ecocycleapp.components.TabBar
 import br.com.ecocycle.ecocycleapp.ui.theme.EcoCycleAppTheme
-
-@Composable
-fun LogoEcoCycle() {
-    Image(
-        painter = painterResource(id = R.drawable.ecocyclelogo),
-        contentDescription = "logo"
-    )
-}
-
-@Composable
-fun getColorId(@ColorRes colorResId: Int): Color {
-    val context = LocalContext.current
-    return Color(context.getColor(colorResId))
-}
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -49,5 +29,5 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 fun RenderScreen() {
-    TabBar()
+
 }
