+++ app/src/main/java/br/com/ecocycle/ecocycleapp/components/BotaoPadrao.kt
@@ -35,10 +35,9 @@ fun BotaoPadrao(
                 id = contentColor!!
             )
         ),
-        elevation = ButtonDefaults.buttonElevation(defaultElevation = 8.dp),
+        elevation = ButtonDefaults.buttonElevation(defaultElevation = 20.dp),
         border = BorderStroke(1.dp, colorResource(id = R.color.transparent))
     ) {
         Text(text = text, fontSize = contentFontSize.sp)
     }
 }
-
