+++ app/src/main/java/br/com/ecocycle/ecocycleapp/screens/reciclagem/Reciclagem.kt
@@ -1,20 +1,2 @@
 package br.com.ecocycle.ecocycleapp.screens.reciclagem
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-
-@Composable
-fun ReciclagemScreen(){
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
-    ) {
-        Text(text = "ReciclagemScreen")
-    }
-}
